Rename SwiftPMConfig to Workspace.Configuration
