+++ Sources/Workspace/CMakeLists.txt
@@ -15,10 +15,10 @@ add_library(Workspace
   ManagedDependency.swift
   ResolvedFileWatcher.swift
   ResolverPrecomputationProvider.swift
-  SwiftPMConfig.swift
   ToolsVersionWriter.swift
   UserToolchain.swift
   Workspace.swift
+  WorkspaceConfiguration.swift
   WorkspaceState.swift
   WindowsToolchainInfo.swift)
 target_link_libraries(Workspace PUBLIC
