+++ Sources/Workspace/Workspace.swift
@@ -166,7 +166,7 @@ public class Workspace {
     public let dataPath: AbsolutePath
 
     /// The swiftpm config.
-    fileprivate let config: SwiftPMConfig
+    fileprivate let config: Workspace.Configuration
 
     /// The current persisted state of the workspace.
     public let state: WorkspaceState
@@ -254,7 +254,7 @@ public class Workspace {
         currentToolsVersion: ToolsVersion = ToolsVersion.currentToolsVersion,
         toolsVersionLoader: ToolsVersionLoaderProtocol = ToolsVersionLoader(),
         delegate: WorkspaceDelegate? = nil,
-        config: SwiftPMConfig = SwiftPMConfig(),
+        config: Workspace.Configuration = Workspace.Configuration(),
         fileSystem: FileSystem = localFileSystem,
         repositoryProvider: RepositoryProvider = GitRepositoryProvider(),
         downloader: Downloader = FoundationDownloader(),
